improved find-sync-peer protocol
tries to fix situation where some clients assumed to be alone (timeout expired)
and thus set themselves 'ready', so they propagated their wrong states to other
peers.

if there are peers around who already found a sync peer, but are not yet ready
the timeout will be increased and if it expires then, a new request is sent. Only
if there is no response at all, it is assumed that no other peers around and only
then it is assumed valid to set itself to 'ready'.