small changes to allow preset management
The goal is not to send preset data back to the requester in order to avoid
reloading the preset on itself whenever it is requested.

The solution is to use a special sender ID 'g' that gets transformed to 'p'
by [netpd-r]. Presets are requested on 'g' and sent to 'p'.