make netpd-bus-order symmetric by allowing responses in both directions
CAVEAT: If responses trigger responses themselves, dulplication of messages
still occurs. In order to avoid that, netpd-bus-order would have to manage
a list of messages to be blocked to make it work for recursive use cases. Since
initialization is idem-potent, sending the same message doesn't cause any harm,
so we do not address that problem now.