WIP: rewrite unpatch-preset-load
Previously, the code was very complex and nested. This is an attempt to fix
the structure and to restore the abililty to read the old text-based format
. Also, the messages are broadcasted, so that all clients receive them.

TODO: the client sending the preset doesn't set the visible elements of GUIs
because remote_origin is not true for the incoming messages. This to be overriden
somehow. Also, the loaded instruments look for a sync peer which they shouldn't
do when loading a preset.